Create README.md
Verelogic provide contract and permanent resourcing services across a broad range of sectors. We understand and appreciate that every company is unique, has its own budget, ways of working and culture. 
We take the time to understand your business, each vacancy in its own right and the personality fit for you as a company. As well as understanding your business, we will interview, understand and inform every candidate about your company, processes and culture. This way, every candidate submitted is fully informed, relevant and right for the role, your company and the working environment/culture.

For that reason, we are committed to providing you with: 

• An honest, consultative and professional approach to recruitment
• Building strong, long-term relationships with business partners across the UK
• Over 30 years combined experience in the recruitment industry
• Providing a consistent service - day in, day out
• Using the latest techniques in acquiring the best candidates the market has to offer
• Value for money is assured 
• Finding people, you'll want to work with – matching skills, culture and personality

Get in touch today to discuss your requirements, 07740280875 / 01793411124 or drop me an email sdunn@verelogic.co.uk